refactor s3dis dataset
